+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.10] - 2022-10-05
 ### Added
 - Warmup-2 stringBits
+- Warmup-2 stringSplosion
 
 ## [1.0.9] - 2022-10-03
 ### Added